Toggle /home not to be on a separate partition
POO#15242
POO#15406

If `/home` is kept on a separate partition (the default), Leap 42.1 YaST
installer choose root partition so small that update to latest 42.1
MaintUpdates and then to Leap 43.1 snapshot fails due to low disk space.
This is avoided if `/home` is not a distinct partition due to
`TOGGLEHOME=1`.

Verification run: http://assam.suse.cz/tests/4669