Add new tag to reuse inst-overview needles
During execution on slow system loading of overview may take time,
whereas we start pressing tab key until get to the software section in
select packages testcase. Software label is in the top, and if we
press tab couple of times before, we won't be able to reach software
section in 20 presses for some installations. We use "inst-overview" tag
in other testcases, and with x11 it may not match. On opensuse we can
simply reuse existing tags, as they verify overview section and hence
confirm that it is loaded.

See poo#12322.