Needles for stabilizing yast2_ui tests
Now I've changed approach to intoroducing changes to multiple tests in
one run according to my findings on the runs. Process reminds a bit
pluging holes in a sinking boat with fingers. On top of that chances to
spot new potential issue are higher when tests are executed as a suite
and not separately.
See poo#19922 and all blocking tasks.