Add needles to accept not latest java version
When we test java on mozilla, occasionally, we have not latest version
installed. It also is not yet in official openSUSE repos, hence we
cannot simply update java. But as far as aim of the test is to
verify that java version is there, we can ignore this warning, as long as
java version is detected. This commit includes a needle to match screen
when not latest java version is detected.

See poo#20632