Delete unsafe needle replaced by addon_products-addon-selection-text-20170805
The previous needle was not checking that the screen is actually ready to
accept input. For example
https://openqa.opensuse.org/tests/459912#step/addon_products/4 shows the "Add"
key selected which was caused by the action before entering the dialog to enter
a URL. Then the screen updates with the added repository and the "Next" gets
highlighted. Only then the screen is ready to accept new input.

See https://openqa.opensuse.org/tests/459912

Related progress issue: https://progress.opensuse.org/issues/5830