Add needles of gdm to fix poo#33484
The default user of gdm will be selected for the scenario of reboot or
first boot, we can just press enter to login. But for the scenario of
log out and re-login we have to use assert_and_click to work around
bsc#1086425.

These needles will be used to check if the keyboard focus of gdm user
is lost.

  see also: poo#33484, bsc#1086425