Workaround typing issue on user_settings
Sporadic typing issues appear when the openQA worker is under heavy
workload. The workaround detect if the fullname is received by the SUT
as expected and if not it retypes the fullname.

- ignore-full-name should just check that the password is typed
entirely, but ignoring fullname since it would not match in case of
mistyping

- expected-typefaces will check if the fullname is received by the SUT
as expected to decide if a retype is necessary.